+++ FinanceApp.Shared/Dto/FGTS/CreateOrUpdateFGTS.cs
@@ -5,8 +5,8 @@ namespace FinanceApp.Shared.Dto.FGTS
     public class CreateOrUpdateFGTS : UpdateDto
     {
 
-        public decimal CurrentBalance { get; set; }
-        public decimal MonthlyGrossIncome { get; set; }
+        public double CurrentBalance { get; set; }
+        public double MonthlyGrossIncome { get; set; }
         public bool AnniversaryWithdraw { get; set; }
 
     }
