+++ FinanceApp.Shared/Dto/FGTS/FGTSDto.cs
@@ -4,8 +4,8 @@ namespace FinanceApp.Shared.Dto.FGTS
 {
     public class FGTSDto : StandardDto
     {
-        public decimal CurrentBalance { get; set; }
-        public decimal MonthlyGrossIncome { get; set; }
+        public double CurrentBalance { get; set; }
+        public double MonthlyGrossIncome { get; set; }
         public bool AnniversaryWithdraw { get; set; }
     }
 }
