+++ FinanceApp.Shared/Models/UserTables/CurrentBalance.cs
@@ -4,8 +4,8 @@ namespace FinanceApp.Shared.Models.UserTables
 {
     public class CurrentBalance : UserTable
     {
-        public decimal Value { get; set; }
-        public decimal? PercentageCdi { get; set; }
+        public double Value { get; set; }
+        public double? PercentageCdi { get; set; }
         public bool UpdateValueWithCdiIndex { get; set; }
     }
 }
