+++ FinanceApp.Shared/Models/UserTables/TreasuryBond.cs
@@ -6,11 +6,11 @@ namespace FinanceApp.Shared.Models.UserTables
     public class TreasuryBond : UserTable
     {
         public ETreasuryBond Type { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public DateTime ExpirationDate { get; set; }
         public DateTime InvestmentDate { get; set; }
         public EOperation Operation { get; set; }
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
 
     }
 }
